+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.11] - 2026-04-08
+
+### Fixed
+- Fix language detection to properly follow system setting (use Locale.preferredLanguages)
+- Fix settings window crash on close (EXC_BAD_ACCESS due to over-release)
+
 ## [1.0.10] - 2026-04-08
 
 ### Improved